Modified   Project/BinarySerializerManaged/BufferStream.cs Modified   Project/NetworkingManaged/Client.cs Modified   Project/NetworkingManaged/ClientSocket.cs Modified   Project/NetworkingManaged/Constants.cs Modified   Project/NetworkingManaged/ServerSocket.cs Modified   Project/Tests/ClientNetworkingManagedTest/Program.cs Modified   Project/Tests/ServerNetworkingManagedTest/Program.cs Modified   Project/Binaries/BinarySerializerManaged.dll Modified   Project/Binaries/CommonManaged.dll Modified   Project/Binaries/NetworkingManaged.NetCore.dll Modified   Project/Binaries/NetworkingManaged.NetFramework.dll

+++ Project/BinarySerializerManaged/BufferStream.cs
@@ -91,7 +91,7 @@ namespace GameFramework.BinarySerializer
 		{
 			byte[] data = new byte[sizeof(double)];
 			ReadBytes(data, 0, sizeof(double));
-			return BitConverter.ToInt64(data, 0);
+			return BitConverter.ToDouble(data, 0);
 		}
 
 		public string ReadString()
